+++ plone/app/iterate/__init__.py
@@ -22,7 +22,25 @@
 """
 """
 
+import logging
 from zope.i18nmessageid import MessageFactory
-PloneMessageFactory = MessageFactory('plone')
+from plone.app.iterate import permissions  # noqa
 
-from plone.app.iterate import permissions
+PloneMessageFactory = MessageFactory('plone')
+logger = logging.getLogger('plone.app.iterate')
+
+
+try:
+    import plone.app.relationfield  # noqa
+except ImportError:
+    logger.warn('Dexterity support for iterate is not available. '
+                'You must install plone.app.relationfield')
+
+
+try:
+    import plone.app.stagingbehavior  # noqa
+    logger.error('plone.app.stagingbehavior should NOT be installed with this version '
+                 'of plone.app.iterate. You may experience problems running this configuration. '
+                 'plone.app.iterate now has dexterity suport built-in.')
+except ImportError:
+    pass