+++ plone/app/iterate/browser/diff.py
@@ -6,30 +6,26 @@ from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
 
 from plone.app.iterate.interfaces import IWorkingCopy, IBaseline
-from plone.app.iterate.relation import WorkingCopyRelation
+from plone.app.iterate.interfaces import ICheckinCheckoutPolicy
 
-class DiffView( BrowserView ):
 
-    def __init__( self, context, request ):
-        self.context = context
-        self.request = request
-        if IBaseline.providedBy( self.context ):
-            self.baseline = context
-            self.working_copy = context.getBackReferences( WorkingCopyRelation.relationship )[0]
-        elif IWorkingCopy.providedBy( self.context ):
-            self.working_copy = context
-            self.baseline = context.getReferences( WorkingCopyRelation.relationship )[0]
+class DiffView(BrowserView):
+
+    def __call__(self):
+        policy = ICheckinCheckoutPolicy(self.context)
+        if IBaseline.providedBy(self.context):
+            self.baseline = self.context
+            self.working_copy = policy.getWorkingCopy()
+        elif IWorkingCopy.providedBy(self.context):
+            self.working_copy = self.context
+            self.baseline = policy.getBaseline()
         else:
             raise AttributeError("Invalid Context")
+        return self.index()
 
-    def diffs( self ):
+    def diffs(self):
         diff = getToolByName(self.context, 'portal_diff')
-        return diff.createChangeSet( self.baseline,
-                                     self.working_copy,
-                                     id1="Baseline",
-                                     id2="Working Copy" )
-
-
-
-
-
+        return diff.createChangeSet(self.baseline,
+                                    self.working_copy,
+                                    id1="Baseline",
+                                    id2="Working Copy")