+++ plone/app/iterate/dexterity/__init__.py
@@ -1,0 +1,2 @@
+
+ITERATE_RELATION_NAME = 'iterate-working-copy'