+++ CHANGES.rst
@@ -1,8 +1,12 @@
 Changelog
 =========
 
-3.0.2 (unreleased)
+3.1.0 (unreleased)
 ------------------
+
+- merge plone.app.stagingbehavior into plone.app.iterate without the
+  behavior implementation. This is for Plone 5 iterate support
+  [vangheem]
 
 - Don't remove aquisition on object for getToolByName call
   [tomgross]